Create Resources

Create resources via Trello API.

create_resource(
  resource,
  id = NULL,
  path = NULL,
  body = list(name = "New"),
  token = NULL,
  response = c("content", "headers", "status", "response"),
  on.error = c("stop", "warn", "message"),
  verbose = FALSE,
  encode = "json",
  handle = NULL
)

Arguments

resource
Model name, eg. "card".

id
Model id.

path
Path.

body
A named list.

token
An object of class "Trello_API_token", a path to a cache file or NULL.

If a Token, it is passed as is.

If NULL and a cache file called ".httr-oauth" exists, the newest token is read from it. If the file is not found, an error is thrown.

If a character vector of length 1, it will be used as an alternative path to the cache file.

response
Can return "content" (the default), "headers", "status" or the raw "response".

on.error
Whether to "stop", "warn" or "message" on http error.

verbose
Whether to pass httr::verbose to httr::PUT.

encode, handle
Passed to httr::POST.

Details

See Trello API reference for more info about what elements can be included in POST request body.

Examples

if (FALSE) {

# Get token with write access
key = Sys.getenv("MY_TRELLO_KEY")
secret = Sys.getenv("MY_TRELLO_SECRET")

token = get_token("my_app", key = key, secret = secret,
                  scope = c("read", "write"))

# Get board ID
url = "Your board URL"
bid = get_id_board(url, token)

# Get lists on that board, extract ID of the first one
lid = get_board_lists(bid, token)$id[1]

# Content for the new card
payload = list(
  idList = lid,
  name = "A new card",
  desc = "#This card has been created by trelloR",
  pos = "bottom"
)

# Create card and store the response (to capture the ID
# of the newly created resource)
r = create_resource("card", body = payload, token = token)

# Get ID of the new card
r$id
}