Get Data From Trello API

Issues GET requests for Trello API endpoints.

get_model(parent = NULL, child = NULL, id = NULL, token = NULL,
  query = NULL, url = NULL, filter = NULL, limit = 1000,
  response = "content", paging = FALSE, bind.rows = TRUE,
  add.class = TRUE, on.error = "error", handle = NULL,
  verbose = FALSE)

Arguments

parent
Parent resource (e.g. "board")

child
Child resource (e.g. "card")

id
Model ID

token
Secure token, see get_token for how to obtain it; if NULL, it will check for a cached token

query
Key-value pairs which form the query, see GET for details

url
Url for the GET request, use instead of specifying parent, id and child; see GET for details

filter
Defaults to "all" which fetches both open and archived cards or all action types

limit
Defaults to 1000; set to 0 to get everything

response
Can be either "content" (a tbl) or an object of class response

paging
Deprecated, use limit = 0 instead

bind.rows
Deprecated; will always try to bind rows unless response is not "content"

add.class
Assign additional S3 class (defaults to TRUE). The additional class is currently not being used for anything, and might be dropped in the future.

on.error
Whether to stop, fail or message on error

handle
The handle to use with this request (see RETRY)

verbose
Set to TRUE for verbose output

Details

If the request fails, server error messages are reprinted on the console.

Only JSON responses are accepted. fromJSON converts them into flat data.frames or lists, while non-JSON type of response throws an error.

When limit > 1000, paging is used. The ID of the oldest result is retrieved from every page and supplied to the next request as the value of the "before" parameter. Paging will stop after reaching the number of results specified by limit, or when there are no more data to get.

filter and limit are query parameters that can be set individually; you could achieve the same result by using query = list(filter = "filter_value", limit = "limit_value")

See also

Examples

# No authorization is required to access public boards, so there is no need
# to create a token, just the board id:
url = "https://trello.com/b/Pw3EioMM/trellor-r-api-for-trello"
bid = get_id_board(url)

#> 
#> GET URL:
#> https://api.trello.com/1/board/Pw3EioMM?fields=name&limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Fetched 1 result(s)

# Once we have the ID, we can use it to make specific queries:
labels = get_board_labels(bid)                # Get all labels

#> 
#> GET URL:
#> https://api.trello.com/1/board/5c2776ec65c66e24e0763a5a/labels?limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Fetched 6 result(s)

cards = get_board_cards(bid, filter = "all")  # Get all cards, incl. archived

#> 
#> GET URL:
#> https://api.trello.com/1/board/5c2776ec65c66e24e0763a5a/cards?limit=1000&filter=all

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================                                                   |  27%  |                                                                              |======================================================================| 100%

#> Fetched 15 result(s)

# We can also call get_model() directly:
lists = get_model(parent = "board", child = "lists", id = bid)

#> 
#> GET URL:
#> https://api.trello.com/1/board/5c2776ec65c66e24e0763a5a/lists?limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Fetched 3 result(s)

# As with boards, cards can be queried for particular resources:
card10 = cards$id[10]
acts10 = get_card_actions(card10)    # Get all actions performed on that card

#> 
#> GET URL:
#> https://api.trello.com/1/card/5c2776ec65c66e24e0763a76/actions?limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Fetched 3 result(s)

# To specify the number of results, use limit = number. If limit = 0, all
# results will be acquired eventually.

if (FALSE) {
acts_all = get_board_actions(bid, limit = 0)
}

# For private and team boards, you need a secure token:

if (FALSE) {
token = get_token("your_key", "your_secret")
cards_open = get_board_cards(board_id, token, filter = "open")
}