Get Data From Trello API

Fetch resources using Trello API.

get_resource(
  parent = NULL,
  child = NULL,
  id = NULL,
  token = NULL,
  query = NULL,
  url = NULL,
  filter = NULL,
  limit = 1000,
  on.error = c("stop", "warn", "message"),
  retry.times = 3,
  handle = NULL,
  verbose = FALSE,
  response,
  paging,
  bind.rows
)

Arguments

parent
Parent resource, e.g. "board" or NULL.

child
Child resource, eg. "card" or NULL.

id
Resource ID or NULL.

token
An object of class "Trello_API_token", a path to a cache file or NULL.

If a token, it is passed as is.

If NULL and a cache file called ".httr-oauth" exists, the newest token is read from it. If the file is not found, an error is thrown.

If a character vector of length 1, it will be used as an alternative path to the cache file.

query
Named list of key-value pairs, see httr::GET for details.

url
Url for the GET request. Can be NULL if parent is specified, or a combination of parent, child and id is provided.

filter
Defaults to "all" which includes both open and archived cards or all action types, depending on what resource is requested.

limit
Defaults to 1000. Set to Inf (or 0) to get everything.

on.error
Whether to "stop", "warn" or "message" on API error.

retry.times
How many times to re-try when a request fails. Defaults to 3.

handle
The handle to use with this request, see httr::RETRY.

verbose
Set to TRUE for verbose output.

response, paging, bind.rows
Deprecated.

Value

A data frame with API responses.

Request limits

At maximum, the API can retrieve 1000 results in a single call. Setting limit > 1000 will activate paging. When paging is used, the request will be issued repeatedly, retrieving new batch of results each time until the limit is reached or there is nothing else to fetch. Results are fetched chronologically, ie. newest results are retrieved first (eg. newest cards). Use limit = Inf to make sure you get all.

Errors

If the request fails, server error messages are reprinted on the console. Depending on the value of on.error, the request call can throw an error in R (this is the default), or can issue a warning/message. If the latter, the function returns a data frame containing the failed URL, HTTP status and an informative message (produced by the server).

Results

The API returns JSON objects which are parsed using jsonlite::fromJSON. Non-JSON results throw an error, but these should never happen anyway. The result is always a data frame, or a tibble if the package is installed.

Filter

Both filter and limit exist as explicitly defined arguments, but you can ignore them in favor of supplying their values as query parameters, eg. query = list(filter = "filter_value", limit = "limit_value").

See also

Examples

# No authorization is required to access public boards, so there is no need
# to create a token, just the board id:
url = "https://trello.com/b/Pw3EioMM/trellor-r-api-for-trello"
bid = get_id_board(url)

#> 
#> GET URL:
#> https://api.trello.com/1/board/Pw3EioMM?fields=name&limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

# Once we have the ID, we can use it to make specific queries:
labels = get_board_labels(bid)                # Get all labels

#> 
#> GET URL:
#> https://api.trello.com/1/board/5c2776ec65c66e24e0763a5a/labels?limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 
#> Fetched 6 results

#> 
#> Request complete: 6 results.

cards = get_board_cards(bid, filter = "all")  # Get all cards, incl. archived

#> 
#> GET URL:
#> https://api.trello.com/1/board/5c2776ec65c66e24e0763a5a/cards?limit=1000&filter=all

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 
#> Fetched 15 results

#> 
#> Request complete: 15 results.

# We can also call get_resource() directly:
lists = get_resource(parent = "board", child = "lists", id = bid)

#> 
#> GET URL:
#> https://api.trello.com/1/board/5c2776ec65c66e24e0763a5a/lists?limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 
#> Fetched 3 results

#> 
#> Request complete: 3 results.

# As with boards, cards can be queried for particular resources:
card10 = cards$id[10]
acts10 = get_card_actions(card10)    # Get all actions performed on that card

#> 
#> GET URL:
#> https://api.trello.com/1/card/5c2776ec65c66e24e0763a76/actions?limit=1000

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 
#> Fetched 3 results

#> 
#> Request complete: 3 results.

# To specify the number of results, use limit = number. If limit = 0, all
# results will be acquired eventually.

if (FALSE) {
acts_all = get_board_actions(bid, limit = 0)
}

# For private and team boards, you need a secure token:

if (FALSE) {
key = Sys.getenv("MY_TRELLO_KEY")
secret = Sys.getenv("MY_TRELLO_SECRET")

token = get_token("my_app", key = key, secret = secret,
                  scope = c("read", "write"))

cards_open = get_board_cards(board_id, token, filter = "open")
}