Get OAuth1.0 Token

Authorize access to Trello API.

get_token(key, secret, appname = "trello-app", scope = "read",
  expiration = "30days")

Arguments

key
developer key

secret
developer secret

appname
optional app name, defaults to "trello-app"

scope
defaults to "read"; allowed are c("read", "write", "account") or "read,write,account" if you prefer

expiration
defaults to "30days"; also allowed "1hour", "1day" and "never"

Details

To access private data, a secure token is required. In order to create it, you will need your developer credentials ('key' and 'secret') - these can be obtained in the Developer guide after login.

First time you create a token, you will be prompted to confirm the authorization in a browser (you only need to do this once). You will also be offered an option to store the authentication data in your working directory. Keep in mind you have to store your credentials in a secure, non-shared location. get_token uses authentication functions from httr.

See also

oauth_app, oauth_endpoint, oauth1.0_token

Examples

# Source credentials from non-shared location (important!)

# NOT RUN {
source("mykeys.R")
token = get_token(key, secret, scope = "read,write")
# }