Search Trello

Search for resources.

search_resource(
  string,
  resource = "all",
  boards = "mine",
  partial = FALSE,
  query = list(),
  modeltype,
  ...
)

search_cards(string, boards = "mine", partial = FALSE, query = list(), ...)

search_actions(string, boards = "mine", partial = FALSE, query = list(), ...)

search_boards(string, boards = "mine", partial = FALSE, query = list(), ...)

search_members(string, boards = "mine", partial = FALSE, query = list(), ...)

search_teams(string, boards = "mine", partial = FALSE, query = list(), ...)

Arguments

string
Text to search for.

resource
Type of resource to return; typically "cards". Defaults to "all".

boards
Boards to limit the search to - defaults to "mine".

partial
Should partial matching be used? Defaults to FALSE.

query
Name list of additional query parameters; consult search API reference

modeltype
Deprecated, use resource instead.

...
Additional arguments passed to get_resource().

Value

A data frame.

Details

Search can be narrowed down by resource type and will return a single data frame per each type (if anything is found). The value of the resource argument is passed on modelTypes - see other attributes at search API reference.

See also

Examples

if (FALSE) {

search_resource("Anything with this text")
search_cards("A card with this text")

}