Get Data From Trello API

Issues GET requests for Trello API endpoints.

get_model(parent = NULL, child = NULL, id = NULL, token = NULL,
  query = NULL, url = NULL, filter = NULL, limit = 1000,
  paging = FALSE, bind.rows = TRUE, add.class = TRUE)

Arguments

parent
Parent structure (e.g. "board")

child
Child structure (e.g. "card")

id
Model ID

token
Secure token, see get_token for how to obtain it

query
Key-value pairs which form the query, see GET for details

url
Url for the GET request, use instead of specifying parent, id and child; see GET for details

filter
Filter results by this string

limit
Defaults to 1000; set to 0 to get everything

paging
Deprecated, use limit = 0 instead

bind.rows
By default, pages will be combined into one data.frame by bind_rows. Set to FALSE if you want list instead. This is useful on the rare occasion that the JSON response is not formatted correctly and makes bind_rows fail

add.class
Assign additional S3 class (defaults to TRUE)

Details

If the request fails, server error messages are reprinted on the console.

Only JSON responses are accepted. fromJSON converts them into flat data.frames or lists, while non-JSON type of response throws an error.

When limit > 1000, paging is used. The ID of the earliest result is retrieved from every page and supplied to the next request as the value of the "before" parameter. PAging will stop after reaching the number of results specified by limit, or when there are no more data to get. filter and limit are query parameters that can be set individually; you could achieve the same result by using query = list(filter = "filter_value", limit = "limit_value")

See also

GET, fromJSON, get_token, get_id

Examples

# No authorization is required to access public boards. Let's get
# the ID of Trello Development Roadmap board:
url = "https://trello.com/b/nC8QJJoZ/trello-development-roadmap"
bid = get_id_board(url)

#> Request URL:
#> https://api.trello.com/1/board/nC8QJJoZ

#> Request complete

# Once we have the ID, we can use it to make specific queries:
labels = get_board_labels(bid)           # Get all labels

#> Request URL:
#> https://api.trello.com/1/board/4d5ea62fd76aa1136000000c/labels

#> Received 6 results

#> Request complete: 6 results

cards = get_board_cards(bid, limit = 5)  # Get 5 cards

#> Request URL:
#> https://api.trello.com/1/board/4d5ea62fd76aa1136000000c/cards

#> Received 5 results

#> Request complete: 5 results

# We can also call get_model() directly:
lists = get_model(parent = "board", child = "lists", id = bid)

#> Request URL:
#> https://api.trello.com/1/board/4d5ea62fd76aa1136000000c/lists

#> Received 32 results

#> Request complete: 32 results

# As with boards, cards can be queried for particular resources:
card1id = cards$id[1]
card1act = get_card_actions(card1id) # Get all actions performed on that card

#> Request URL:
#> https://api.trello.com/1/card/59c03456e1ea9ffc086f1ca3/actions

#> Response was empty

#> Request complete

#> Warning: Could not assign additional S3 class.

# To retrieve large results, paging might be necessary:

# NOT RUN {
tdr_actions = get_board_actions(bid, paging = TRUE)
# }

# For private boards, you need a secure token:

# NOT RUN {
token = get_token("your_key", "your_secret")
all_open_cards = get_board_cards(board_id, token, filter = "open")
# }